Move related functions into their own files Also noted where functions are called only within the file in which they are declared
